refactor(1015): extract PromptUtils to src/ui/prompt_utils (T-07, Cat E)
Moves the 441-line PromptUtils class (29 static methods) out of
MistHelper.py into a dedicated module, leaving a bare re-export at the
original site. Cross-module MistHelper globals (apisession, DataExporter,
LAST_SELECTED_SITE_ID) are reached via the lazy-mh pattern used in prior
T-extractions.

Adds src/ui/prompt_utils.py to the coverage omit list, matching the
policy for other interactive prompt modules (src/ui/tui.py,
src/input/prompt_client_utils.py) whose behavior is user-input driven
and not amenable to line-coverage measurement.

Also drops unused `Literal` from MistHelper.py's typing import.
